Move ScriptingContext into dedicated header file
